backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
Pillow>=10.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
openai>=1.0.0
google-generativeai>=0.3.0
httpx>=0.25.0
PyMuPDF>=1.23.0